Refactor ShowMigrator. Extract migrations to their own files.
